Generate meaningful COMPILER_ID/REPO_SCHEMA when not on git
Currently if HHVM is built outside of git, the COMPILER_ID
and REPO_SCHEMA values will effectively be constants which
means that schema changes won't be picked up properly.

This diff works around the absence of git by using the system
time (nanosecond scale) to make a "random enough" number.

+++ bin/generate_repo_schema.sh
@@ -3,15 +3,22 @@
 cd ${HPHP_HOME}
 # Generate src/runtime/vm/repo_schema.h
 REPO_TREE=`git log -n1 --pretty=format:%T HEAD`
-REPO_MODS=`git diff --name-only HEAD`
-if [ "${REPO_MODS}" != "" ]; then
-  REPO_TREE=`
-  export GIT_INDEX_FILE=.git-index-$$$$;
-  git read-tree ${REPO_TREE};
-  git update-index --add --remove ${REPO_MODS};
-  git write-tree;
-  rm -f $GIT_INDEX_FILE`
+if [ -z "${REPO_TREE}" ]; then
+  # HipHop has been divorced from git,
+  # define REPO_SCHEMA as current time
+  # to play it safe
+  HHVM_REPO_SCHEMA=`date +%s%N`
+else
+  REPO_MODS=`git diff --name-only HEAD`
+  if [ "${REPO_MODS}" != "" ]; then
+    REPO_TREE=`
+    export GIT_INDEX_FILE=.git-index-$$$$;
+    git read-tree ${REPO_TREE};
+    git update-index --add --remove ${REPO_MODS};
+    git write-tree;
+    rm -f $GIT_INDEX_FILE`
+  fi
+  HHVM_REPO_SCHEMA=`git ls-tree --full-tree ${REPO_TREE} src/ | grep -v src/test | git hash-object --stdin`
 fi
-HHVM_REPO_SCHEMA=`git ls-tree --full-tree ${REPO_TREE} src/ | grep -v src/test | git hash-object --stdin`
-echo "#define REPO_SCHEMA \"${HHVM_REPO_SCHEMA}\"" > src/runtime/vm/repo_schema.h
 
+echo "#define REPO_SCHEMA \"${HHVM_REPO_SCHEMA}\"" > src/runtime/vm/repo_schema.h
